% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pokemon.R
\name{get_pokemon}
\alias{get_pokemon}
\title{Get Pokémon Information}
\usage{
get_pokemon(names)
}
\arguments{
\item{names}{A character vector of Pokémon names.}
}
\value{
A data frame with one row per Pokémon.
}
\description{
Retrieve the main data for one or more Pokémon.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
get_pokemon(c("pikachu", "fuecoco"))
\dontshow{\}) # examplesIf}
}
