% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_matchup.R
\name{type_matchup}
\alias{type_matchup}
\title{Type Matchup}
\usage{
type_matchup(name1, name2)
}
\arguments{
\item{name1}{The name of the first Pokémon.}

\item{name2}{The name of the second Pokémon.}
}
\value{
A data frame with one row: \code{attack} is the multiplier when
the first attacks the second, and \code{defense} when the second attacks
the first.
}
\description{
Compare the types of two Pokémon.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
type_matchup("pikachu", "fuecoco")
\dontshow{\}) # examplesIf}
}
