% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species.R
\name{get_species}
\alias{get_species}
\title{Get Species Information}
\usage{
get_species(names)
}
\arguments{
\item{names}{A character vector of Pokémon species names.}
}
\value{
A data frame with columns: name, capture_rate, growth_rate,
rarity_level, gender_rate, evolves_from_species, evolution_chain and
generation.
}
\description{
Retrieve species-level information for one or more Pokémon.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
get_species("Fuecoco")
\dontshow{\}) # examplesIf}
}
