% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\docType{class}
\name{pokemon-class}
\alias{pokemon-class}
\alias{pokemon}
\title{A Pokémon Object}
\description{
An RC class representing a Pokémon. Creating an object requires an internet connection.
}
\section{Fields}{

\describe{
\item{\code{name}}{The name of the Pokémon.}

\item{\code{id}}{The Pokédex ID.}

\item{\code{height}}{The height in decimetres.}

\item{\code{weight}}{The weight in hectograms.}

\item{\code{base_experience}}{The base experience.}

\item{\code{types}}{A character vector of types.}

\item{\code{stats}}{A data frame of base stats.}

\item{\code{abilities}}{A character vector of abilities.}

\item{\code{species}}{The species name.}

\item{\code{evolution_chain}}{The evolution chain.}

\item{\code{moves}}{A character vector of moves.}

\item{\code{sprites}}{A list of sprite URLs.}

\item{\code{cries}}{A list of cry URLs.}

\item{\code{held_items}}{A data frame of held items.}

\item{\code{location_area_encounters}}{The location area encounters URL.}

\item{\code{is_legendary}}{Logical. Is it legendary?}

\item{\code{is_mythical}}{Logical. Is it mythical?}

\item{\code{rarity}}{The rarity level (our own definition).}
}}


\section{Methods}{

\describe{
\item{\code{data()}}{Return the main data frame.}
\item{\code{types()}}{Return the types.}
\item{\code{stats()}}{Return the stats.}
\item{\code{abilities()}}{Return the abilities.}
\item{\code{moves()}}{Return the moves.}
}
}

\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
p <- pokemon$new("bulbasaur")
p$data()
\dontshow{\}) # examplesIf}
}
